+++ pxtblocks/blocklyloader.ts
@@ -2,12 +2,6 @@
 /// <reference path="../built/pxtlib.d.ts" />
 
 namespace pxt.blocks {
-    export enum CategoryMode {
-        All,
-        None,
-        Basic
-    }
-
     const typeDefaults: Map<{ field: string, block: string, defaultValue: string }> = {
         "string": {
             field: "TEXT",
@@ -525,61 +519,6 @@ namespace pxt.blocks {
         }
     }
 
-    let toolboxStyle: HTMLStyleElement;
-    let toolboxStyleBuffer: string = '';
-    export function appendToolboxIconCss(className: string, i: string): void {
-        if (toolboxStyleBuffer.indexOf(className) > -1) return;
-
-        if (i.length === 1) {
-            const icon = Util.unicodeToChar(i);
-            toolboxStyleBuffer += `
-                .blocklyTreeIcon.${className}::before {
-                    content: "${icon}";
-                }
-            `;
-        }
-        else {
-            toolboxStyleBuffer += `
-                .blocklyTreeIcon.${className} {
-                    background-image: url("${Util.pathJoin(pxt.webConfig.commitCdnUrl, encodeURI(i))}")!important;
-                    width: 30px;
-                    height: 100%;
-                    background-size: 20px !important;
-                    background-repeat: no-repeat !important;
-                    background-position: 50% 50% !important;
-                }
-            `;
-        }
-    }
-
-    export function injectToolboxIconCss(): void {
-        if (!toolboxStyle) {
-            toolboxStyle = document.createElement('style');
-            toolboxStyle.id = "blocklyToolboxIcons";
-            toolboxStyle.type = 'text/css';
-            let head = document.head || document.getElementsByTagName('head')[0];
-            head.appendChild(toolboxStyle);
-        }
-
-        if (toolboxStyle.sheet) {
-            toolboxStyle.textContent = toolboxStyleBuffer + namespaceStyleBuffer;
-        } else {
-            toolboxStyle.appendChild(document.createTextNode(toolboxStyleBuffer + namespaceStyleBuffer));
-        }
-    }
-
-    let namespaceStyleBuffer: string = '';
-    export function appendNamespaceCss(namespace: string, color: string) {
-        const ns = namespace.toLowerCase();
-        color = color || '#dddddd'; // Default toolbox color
-        if (namespaceStyleBuffer.indexOf(ns) > -1) return;
-        namespaceStyleBuffer += `
-            span.docs.${ns} {
-                background-color: ${color} !important;
-                border-color: ${Blockly.PXTUtils.fadeColour(color, 0.2, true)} !important;
-            }
-        `;
-    }
 
     function getChildCategories(parent: Element) {
         const elements = parent.getElementsByTagName("category");
